Split out Tokens into their own source file.
