+++ src/lib.rs
@@ -45,6 +45,7 @@ pub use will_options::*;        //{WillOptions, WillOptionsBuilder};
 pub use ssl_options::*;         //{SslOptions, SslOptionsBuilder};
 pub use disconnect_options::*;  //{DisconnectOptions, DisconnectOptionsBuilder};
 pub use message::*;             //{Message, MessageBuilder};
+pub use token::*;               //{Token}
 pub use topic::*;               //{Topic}
 pub use client_persistence::*;
 pub use errors::*;              //{MqttResult, MqttError, ErrorKind};
@@ -75,6 +76,9 @@ pub mod disconnect_options;
 
 /// The message object
 pub mod message;
+
+/// Tokens to monitor asynchronous operations
+pub mod token;
 
 /// Options for creating topic objects that are associated with a
 /// particular server.
